+++ include/YODA/Scatter1D.h
@@ -116,7 +116,7 @@ namespace YODA {
 
 
     /// Dimension of this data object
-    int dim() const { return 1; }
+    size_t dim() const { return 1; }
 
 
     /// @name Modifiers
