+++ include/YODA/Scatter2D.h
@@ -126,7 +126,7 @@ namespace YODA {
 
 
     /// Dimension of this data object
-    int dim() const { return 2; }
+    size_t dim() const { return 2; }
 
 
     /// @name Modifiers
